# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SpecForge API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type TechDebtRiskSeverity =
  (typeof TechDebtRiskSeverity)[keyof typeof TechDebtRiskSeverity];

export const TechDebtRiskSeverity = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface TechDebtRisk {
  title: string;
  severity: TechDebtRiskSeverity;
  description: string;
}

export type SpecSpecType = (typeof SpecSpecType)[keyof typeof SpecSpecType];

export const SpecSpecType = {
  system_design: "system_design",
  api_design: "api_design",
  database_schema: "database_schema",
  feature_spec: "feature_spec",
} as const;

export type SpecInputType = (typeof SpecInputType)[keyof typeof SpecInputType];

export const SpecInputType = {
  github_url: "github_url",
  description: "description",
} as const;

export type SpecStatus = (typeof SpecStatus)[keyof typeof SpecStatus];

export const SpecStatus = {
  pending: "pending",
  generating: "generating",
  completed: "completed",
  failed: "failed",
} as const;

export interface Spec {
  id: number;
  title: string;
  specType: SpecSpecType;
  inputType: SpecInputType;
  inputValue: string;
  content: string;
  status: SpecStatus;
  complexityScore?: number | null;
  techDebtRisks?: TechDebtRisk[] | null;
  complexitySummary?: string | null;
  mermaidDiagram?: string | null;
  conversationId?: number | null;
  shareToken?: string | null;
  viewCount: number;
  lastSyncedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ShareSpecResponse {
  shareToken: string;
  shareUrl: string;
  viewCount: number;
}

export type CreateSpecBodySpecType =
  (typeof CreateSpecBodySpecType)[keyof typeof CreateSpecBodySpecType];

export const CreateSpecBodySpecType = {
  system_design: "system_design",
  api_design: "api_design",
  database_schema: "database_schema",
  feature_spec: "feature_spec",
} as const;

export type CreateSpecBodyInputType =
  (typeof CreateSpecBodyInputType)[keyof typeof CreateSpecBodyInputType];

export const CreateSpecBodyInputType = {
  github_url: "github_url",
  description: "description",
} as const;

export interface CreateSpecBody {
  specType: CreateSpecBodySpecType;
  inputType: CreateSpecBodyInputType;
  inputValue: string;
  title: string;
}

export type RecentSpecsResponseByType = { [key: string]: number };

export interface RecentSpecsResponse {
  specs: Spec[];
  totalCount: number;
  byType: RecentSpecsResponseByType;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export interface SendAnthropicMessageBody {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
}

export interface GetCurrentAuthUserResponse {
  isAuthenticated: boolean;
  user?: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  state: string;
  nonce?: string;
  redirect_uri: string;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export interface SyncSpecResponse {
  id: number;
  status: string;
  lastSyncedAt?: string | null;
}

export interface WebhookConfigResponse {
  webhookUrl: string;
  secret: string;
  instructions: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface SpecPresenceViewer {
  sessionId: string;
  name: string;
  color: string;
  joinedAt: string;
}

export type SpecInsightsResponseOverallHealth =
  (typeof SpecInsightsResponseOverallHealth)[keyof typeof SpecInsightsResponseOverallHealth];

export const SpecInsightsResponseOverallHealth = {
  excellent: "excellent",
  good: "good",
  fair: "fair",
  poor: "poor",
} as const;

export interface SpecInsightsResponse {
  completeness: number;
  overallHealth: SpecInsightsResponseOverallHealth;
  missingAreas: string[];
  strengthAreas: string[];
  suggestions: string[];
  estimatedImplementationDays: number;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  sync_complete: "sync_complete",
  sync_failed: "sync_failed",
  share_viewed: "share_viewed",
} as const;

export interface Notification {
  id: number;
  userId: string;
  type: NotificationType;
  title: string;
  message: string;
  specId?: number | null;
  read: boolean;
  createdAt: string;
}

export interface NotificationListResponse {
  notifications: Notification[];
  unreadCount: number;
}

export type LoginParams = {
  returnTo?: string;
};

export type GithubWebhook200 = {
  processed: number;
};