# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SpecForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /specs/{id}/insights:
    post:
      operationId: analyzeSpec
      tags: [specs]
      summary: Get AI-powered health insights and analysis for a spec
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Spec health insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecInsightsResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /notifications/{id}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    SpecInsightsResponse:
      type: object
      properties:
        completeness:
          type: integer
        overallHealth:
          type: string
          enum: [excellent, good, fair, poor]
        missingAreas:
          type: array
          items:
            type: string
        strengthAreas:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        estimatedImplementationDays:
          type: integer
      required:
        - completeness
        - overallHealth
        - missingAreas
        - strengthAreas
        - suggestions
        - estimatedImplementationDays
...